+++ ManejoPresupuesto/ManejoPresupuesto/Models/ReporteTransaccionesDetalladas.cs
@@ -5,17 +5,17 @@
         public DateTime FechaInicio { get; set; }
         public DateTime FechaFin { get; set; }
         public IEnumerable<TransaccionesPorFecha> TransaccionesAgrupadas { get; set; }
-        public int BalanceDepositos => TransaccionesAgrupadas.Sum(x=> x.BalanceDepositos);
-        public int BalanceRetiros => TransaccionesAgrupadas.Sum(x => x.BalanceRetiros);
-        public int Total => BalanceDepositos - BalanceRetiros;
+        public decimal BalanceDepositos => TransaccionesAgrupadas.Sum(x=> x.BalanceDepositos);
+        public decimal BalanceRetiros => TransaccionesAgrupadas.Sum(x => x.BalanceRetiros);
+        public decimal Total => BalanceDepositos - BalanceRetiros;
         public class TransaccionesPorFecha
         { 
             public DateTime FechaTransaccion { get; set; }
             public IEnumerable<Transaccion> Transacciones { get; set; }
-            public int BalanceDepositos => 
+            public decimal BalanceDepositos => 
                 Transacciones.Where(x => x.TipoOperacionId == TipoOperacion.Ingreso)
                 .Sum(x => x.Monto);
-            public int BalanceRetiros =>
+            public decimal BalanceRetiros =>
                   Transacciones.Where(x => x.TipoOperacionId == TipoOperacion.Gasto)
                 .Sum(x => x.Monto);
         }
