+++ ManejoPresupuesto/ManejoPresupuesto/Models/ResultadoObtenerPorSemana.cs
@@ -3,10 +3,10 @@
     public class ResultadoObtenerPorSemana
     {
         public int Semana { get; set; }
-        public int Monto { get; set; }
+        public decimal Monto { get; set; }
         public TipoOperacion TipoOperacionId { get; set; }
-        public int Ingresos { get; set; }
-        public int Gastos { get; set; }
+        public decimal Ingresos { get; set; }
+        public decimal Gastos { get; set; }
 
         public DateTime FechaInicio { get; set; }
         public DateTime FechaFin { get; set; }
